# fruit-store/src/test/resources/openapi.yaml
---
openapi: 3.0.3
info:
  title: Generated API
  version: "1.0"
paths:
  /:
    get:
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                $ref: '#/components/schemas/TemplateInstance'
  /fruits:
    get:
      operationId: listFruits
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetFruit'
    post:
      operationId: addFruit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fruit'
      responses:
        "200":
          description: OK
  /fruits/price/{id}:
    get:
      operationId: getPriceUpdate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /fruits/{id}:
    get:
      operationId: getFruitById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
    delete:
      operationId: deleteFruit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
components:
  schemas:
    TemplateInstance:
      type: object
    Category:
      type: object
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
    BigDecimal:
      type: number
    Status:
      enum:
      - AVAILABLE
      - PENDING
      - SOLD
      type: string
    Fruit:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/Category'
        description:
          type: string
        id:
          format: int64
          type: integer
        name:
          type: string
        price:
          $ref: '#/components/schemas/BigDecimal'
        status:
          $ref: '#/components/schemas/Status'
        tags:
          type: array
          items:
            type: string
    SetFruit:
      uniqueItems: true
      type: array
      items:
        $ref: '#/components/schemas/Fruit'